requirements.txt
numpy==1.13.1
pyqtgraph==0.10.0
pyserial==3.4
